Move non-inline functions out of headers
With this in the header a `multiple definition of `to_simulator'` error is emitted by the linker if two different compilation units include this header.

The alternative was to mark these functions as `inline`, but discussion on github favored not doing this.
