+++ cocotb/share/include/cocotb_utils.h
@@ -34,9 +34,6 @@
 extern "C" {
 #endif
 
-#include <stdio.h>
-#include <stdlib.h>
-
 #define xstr(a) str(a)
 #define str(a) #a
 
@@ -45,24 +42,8 @@ extern void* utils_dyn_sym(void *handle, const char* sym_name);
 
 extern int is_python_context;
 
-void to_python(void) {
-    if (is_python_context) {
-        fprintf(stderr, "FATAL: We are calling up again\n");
-        exit(1);
-    }
-    ++is_python_context;
-    //fprintf(stderr, "INFO: Calling up to python %d\n", is_python_context);
-}
-
-void to_simulator(void) {
-    if (!is_python_context) {
-        fprintf(stderr, "FATAL: We have returned twice from python\n");
-        exit(1);
-    }
-
-    --is_python_context;
-    //fprintf(stderr, "INFO: Returning back to simulator %d\n", is_python_context);
-}
+void to_python(void);
+void to_simulator(void);
 
 #ifdef __cplusplus
 }
