+++ cocotb/share/lib/utils/cocotb_utils.c
@@ -29,6 +29,7 @@
 
 #include <cocotb_utils.h>
 #include <stdio.h>
+#include <stdlib.h>
 
 #if defined(__linux__) || defined(__APPLE__)
 #include <dlfcn.h>
@@ -38,6 +39,25 @@
 
 // Tracks if we are in the context of Python or Simulator
 int is_python_context = 0;
+
+void to_python(void) {
+    if (is_python_context) {
+        fprintf(stderr, "FATAL: We are calling up again\n");
+        exit(1);
+    }
+    ++is_python_context;
+    //fprintf(stderr, "INFO: Calling up to python %d\n", is_python_context);
+}
+
+void to_simulator(void) {
+    if (!is_python_context) {
+        fprintf(stderr, "FATAL: We have returned twice from python\n");
+        exit(1);
+    }
+
+    --is_python_context;
+    //fprintf(stderr, "INFO: Returning back to simulator %d\n", is_python_context);
+}
 
 void* utils_dyn_open(const char* lib_name)
 {
